Add SQLDelight support

+++ shared/src/commonMain/kotlin/com/mecofarid/trending/features/trending/domain/model/Trending.kt
@@ -1,7 +1,7 @@
 package com.mecofarid.trending.features.trending.domain.model
 
 data class Trending(
-    val trendingId: Int,
+    val trendingId: Long,
     val name: String,
     val language: String?,
     val stargazersCount: Long,
